+++ scylla/src/connection/mod.rs
@@ -1,14 +1,15 @@
 mod connection;
 mod connection_pool;
 
+pub use crate::transport::errors::TranslationError;
 use connection::ErrorReceiver;
-pub use connection::SelfIdentity;
 #[cfg(feature = "ssl")]
 pub(crate) use connection::SslConfig;
 pub(crate) use connection::{
     open_connection, Connection, ConnectionConfig, NonErrorQueryResponse, QueryResponse,
     VerifiedKeyspaceName,
 };
+pub use connection::{AddressTranslator, SelfIdentity};
 
 pub use connection_pool::PoolSize;
 pub(crate) use connection_pool::{NodeConnectionPool, PoolConfig};
